[refactor] Refactor C++ aot test to accommodate multiple backends
[ghstack-poisoned]
