+++ tests/cpp/aot/python_scripts/sph_aot.py
